아라사카 동그라미 퍼센트 증가
