+++ Assets/Resources/MusicData.json
@@ -1,1 +1,1 @@
-{"songs":[{"title":"Stylish Rock Beat Trailer","artist":"ComaStudio","cover_image_path":"Image/sumikko","audio_file_path":"Audio/stylish-rock-beat-trailer-ComaStudio","score":0.0,"rank":"F","percentage":0.0},{"title":"Do You Want To Build A Snowman","artist":"Kristen Bell","cover_image_path":"Image/Do You Want To Build A Snowman","audio_file_path":"Audio/Kristen Bell-Do You Want To Build A Snowman","score":0.0,"rank":"","percentage":0.0},{"title":"Happy","artist":"Pharrell Williams","cover_image_path":"Image/happy","audio_file_path":"Audio/Pharrell Williams - 01 - Happy","score":0.0,"rank":"","percentage":0.0},{"title":"Give Love","artist":"AKMU","cover_image_path":"Image/Give Love","audio_file_path":"Audio/AKMU - 01 - Give Love","score":0.0,"rank":"","percentage":0.0},{"title":"RE BYE","artist":"AKMU","cover_image_path":"Image/RE BYE","audio_file_path":"Audio/AKMU-01-RE-BYE-128","score":0.0,"rank":"","percentage":0.0},{"title":"YOU KNOW ME","artist":"AKMU","cover_image_path":"Image/YOU KNOW ME","audio_file_path":"Audio/AKMU-06-YOU KNOW ME","score":0.0,"rank":"","percentage":0.0},{"title":"Run","artist":"EXO","cover_image_path":"Image/Run","audio_file_path":"Audio/EXO-K - 04 - Run","score":0.0,"rank":"","percentage":0.0}]}
+{"songs":[{"title":"Stylish Rock Beat Trailer","artist":"ComaStudio","cover_image_path":"Image/sumikko","audio_file_path":"Audio/stylish-rock-beat-trailer-ComaStudio","score":0.0,"rank":"F","percentage":0.0},{"title":"Do You Want To Build A Snowman","artist":"Kristen Bell","cover_image_path":"Image/Do You Want To Build A Snowman","audio_file_path":"Audio/Kristen Bell-Do You Want To Build A Snowman","score":0.0,"rank":"","percentage":0.0},{"title":"Happy","artist":"Pharrell Williams","cover_image_path":"Image/happy","audio_file_path":"Audio/Pharrell Williams - 01 - Happy","score":0.0,"rank":"","percentage":0.0},{"title":"Give Love","artist":"AKMU","cover_image_path":"Image/Give Love","audio_file_path":"Audio/AKMU - 01 - Give Love","score":0.0,"rank":"","percentage":80.0},{"title":"RE BYE","artist":"AKMU","cover_image_path":"Image/RE BYE","audio_file_path":"Audio/AKMU-01-RE-BYE-128","score":0.0,"rank":"","percentage":90.0},{"title":"YOU KNOW ME","artist":"AKMU","cover_image_path":"Image/YOU KNOW ME","audio_file_path":"Audio/AKMU-06-YOU KNOW ME","score":0.0,"rank":"","percentage":0.0},{"title":"Run","artist":"EXO","cover_image_path":"Image/Run","audio_file_path":"Audio/EXO-K - 04 - Run","score":0.0,"rank":"","percentage":0.0}]}
