+++ Assets/Scenes/TestMainScene.unity
@@ -3270,7 +3270,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 702307152}
-  - component: {fileID: 702307151}
   - component: {fileID: 702307154}
   - component: {fileID: 702307155}
   m_Layer: 0
@@ -3280,34 +3279,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &702307151
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 702307150}
-  m_Enabled: 0
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 937a081a6818df94cbb2f40777e36cdc, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  questPanel: []
-  questPanelPosition:
-  - {fileID: 0}
-  - {fileID: 0}
-  - {fileID: 0}
-  - {fileID: 0}
-  - {fileID: 0}
-  audioClips: []
-  titles: []
-  artists: []
-  albums: []
-  panelPrefab: {fileID: 3421999821491316425, guid: c4bd16c4feac7f54fa7081c077f6b339, type: 3}
-  QuestPanelList: {fileID: 0}
-  MusicDataFile: 
-  title: {fileID: 0}
-  artist: {fileID: 0}
 --- !u!4 &702307152
 Transform:
   m_ObjectHideFlags: 0
@@ -3343,6 +3314,7 @@ MonoBehaviour:
   - {fileID: 973815311}
   panelPrefab: {fileID: 3421999821491316425, guid: c4bd16c4feac7f54fa7081c077f6b339, type: 3}
   QuestPanelList: {fileID: 721440520}
+  isPassed: 0
   title: {fileID: 576105915}
   artist: {fileID: 1271816180}
   score: {fileID: 742141497}
@@ -3443,6 +3415,82 @@ AudioSource:
     m_PreInfinity: 2
     m_PostInfinity: 2
     m_RotationOrder: 4
+--- !u!1 &713015459
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 713015460}
+  - component: {fileID: 713015462}
+  - component: {fileID: 713015461}
+  m_Layer: 5
+  m_Name: Image
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!224 &713015460
+RectTransform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 713015459}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 0.5, y: 0.5, z: 0.05}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 911332143}
+  m_RootOrder: 2
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: 0, y: 0}
+  m_SizeDelta: {x: 1754, y: 1000}
+  m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &713015461
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 713015459}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Material: {fileID: 0}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
+  m_RaycastTarget: 1
+  m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
+  m_Maskable: 1
+  m_OnCullStateChanged:
+    m_PersistentCalls:
+      m_Calls: []
+  m_Sprite: {fileID: 21300000, guid: 6aa286075fe3a7a4fb88f310a5fd03ec, type: 3}
+  m_Type: 0
+  m_PreserveAspect: 0
+  m_FillCenter: 1
+  m_FillMethod: 4
+  m_FillAmount: 1
+  m_FillClockwise: 1
+  m_FillOrigin: 0
+  m_UseSpriteMesh: 0
+  m_PixelsPerUnitMultiplier: 1
+--- !u!222 &713015462
+CanvasRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 713015459}
+  m_CullTransparentMesh: 1
 --- !u!1 &721440520
 GameObject:
   m_ObjectHideFlags: 0
@@ -4534,6 +4582,7 @@ RectTransform:
   m_Children:
   - {fileID: 1649867758}
   - {fileID: 2022892042}
+  - {fileID: 713015460}
   m_Father: {fileID: 1732446651}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -8258,7 +8307,7 @@ MonoBehaviour:
   songSelection: {fileID: 393421562}
   ARASAKAFillAmount: {fileID: 1718334275}
   ARASAKAPercent: {fileID: 440193941}
-  ARASAKAClearPercentage: 0
+  musicManager: {fileID: 702307154}
 --- !u!1 &1882071832
 GameObject:
   m_ObjectHideFlags: 0
@@ -9530,7 +9579,7 @@ GameObject:
   - component: {fileID: 2022892044}
   - component: {fileID: 2022892043}
   m_Layer: 0
-  m_Name: PlayCount (1)
+  m_Name: Under construction
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
