+++ Assets/Scripts/MusicManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 937a081a6818df94cbb2f40777e36cdc
+guid: df79a82f580d88343a2b33bd8d5f1990
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
