--- Assets/Scripts/MusicManager2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: df79a82f580d88343a2b33bd8d5f1990
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
